feat(e1a+p001+drift): cortex crate promotion + P001 SKILL_REGISTRY + boot/checkpoint/drift
E1a: cortex crate promotion (69 erros→0)
- Funcoes bin-specific (generate_via_model, dispatch_expert) movidas do crate p/ bin
- Argmax u16->u32; path rewrites crate:: -> k_nano::
- k-hal xpu.rs u16->u32 token types
- hermes: ToString, k_ai::gguf->cortex::gguf, generate_via_model adicionados ao crate

P001: SKILL_REGISTRY shadow fix
- main.rs private SKILL_REGISTRY removido; register_builtin_skills() no k_nano::SKILL_REGISTRY
- 4 referencias crate::SKILL_REGISTRY -> k_nano::SKILL_REGISTRY

Boot path + Safety I4
- Agency fallback: 2 AgentSpecs quando PACKAGE_HUB vazio
- SafetyAgent: verify_counter + AUDIT_TRAIL.verify() a cada 100 ticks

Checkpoint SelfHeal expandido
- 5 novos campos: heap_start/size, PML4/CR3 addr, driver_state_hash FNV-1a, v=2
- save/restore com diagnostico aprimorado

Drift quick wins: env.rs (is_online movido p/ k_nano), block_dev.rs (UsbMassStorage impl mantido no bin)

Licoes: SKILL_REGISTRY shadow, tipo struct drift, ToString em no_std, return em match arm tipo errado
Ver SESSION_217.md

+++ crates/cortex/src/ngram_spec.rs
@@ -168,11 +168,11 @@ impl NgramSpeculator {
 
 /// Verify draft[1..] against per-position logits from a parallel forward of `drafts`.
 /// Caller must already have verified drafts[0] against the previous-step logits.
-pub fn verify_draft(all_logits: &Tensor, drafts: &[u16]) -> (usize, u16) {
+pub fn verify_draft(all_logits: &Tensor, drafts: &[u16]) -> (usize, u32) {
     let mut accept = 0usize;
     for i in 0..drafts.len().saturating_sub(1) {
         let predicted = argmax_row(all_logits, i);
-        if predicted == drafts[i + 1] {
+        if predicted == drafts[i + 1] as u32 {
             accept += 1;
         } else {
             return (accept, predicted);
